requirements/requirements.development.txt
# debug
django-debug-toolbar
